+++ actix-files/src/lib.rs
@@ -110,7 +110,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_if_modified_since_without_if_none_match() {
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let since = header::HttpDate::from(SystemTime::now().add(Duration::from_secs(60)));
 
         let req = TestRequest::default()
@@ -122,7 +122,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_if_modified_since_without_if_none_match_same() {
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let since = file.last_modified().unwrap();
 
         let req = TestRequest::default()
@@ -134,7 +134,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_if_modified_since_with_if_none_match() {
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let since = header::HttpDate::from(SystemTime::now().add(Duration::from_secs(60)));
 
         let req = TestRequest::default()
@@ -147,7 +147,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_if_unmodified_since() {
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let since = file.last_modified().unwrap();
 
         let req = TestRequest::default()
@@ -159,7 +159,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_if_unmodified_since_failed() {
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let since = header::HttpDate::from(SystemTime::UNIX_EPOCH);
 
         let req = TestRequest::default()
@@ -171,8 +171,8 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_text() {
-        assert!(NamedFile::open_async("test--").await.is_err());
-        let mut file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        assert!(NamedFile::open("test--").is_err());
+        let mut file = NamedFile::open("Cargo.toml").unwrap();
         {
             file.file();
             let _f: &File = &file;
@@ -195,8 +195,8 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_content_disposition() {
-        assert!(NamedFile::open_async("test--").await.is_err());
-        let mut file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        assert!(NamedFile::open("test--").is_err());
+        let mut file = NamedFile::open("Cargo.toml").unwrap();
         {
             file.file();
             let _f: &File = &file;
@@ -212,8 +212,7 @@ mod tests {
             "inline; filename=\"Cargo.toml\""
         );
 
-        let file = NamedFile::open_async("Cargo.toml")
-            .await
+        let file = NamedFile::open("Cargo.toml")
             .unwrap()
             .disable_content_disposition();
         let req = TestRequest::default().to_http_request();
@@ -248,8 +247,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_set_content_type() {
-        let mut file = NamedFile::open_async("Cargo.toml")
-            .await
+        let mut file = NamedFile::open("Cargo.toml")
             .unwrap()
             .set_content_type(mime::TEXT_XML);
         {
@@ -274,7 +272,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_image() {
-        let mut file = NamedFile::open_async("tests/test.png").await.unwrap();
+        let mut file = NamedFile::open("tests/test.png").unwrap();
         {
             file.file();
             let _f: &File = &file;
@@ -297,7 +295,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_javascript() {
-        let file = NamedFile::open_async("tests/test.js").await.unwrap();
+        let file = NamedFile::open("tests/test.js").unwrap();
 
         let req = TestRequest::default().to_http_request();
         let resp = file.respond_to(&req);
@@ -317,8 +315,7 @@ mod tests {
             disposition: DispositionType::Attachment,
             parameters: vec![DispositionParam::Filename(String::from("test.png"))],
         };
-        let mut file = NamedFile::open_async("tests/test.png")
-            .await
+        let mut file = NamedFile::open("tests/test.png")
             .unwrap()
             .set_content_disposition(cd);
         {
@@ -343,7 +340,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_binary() {
-        let mut file = NamedFile::open_async("tests/test.binary").await.unwrap();
+        let mut file = NamedFile::open("tests/test.binary").unwrap();
         {
             file.file();
             let _f: &File = &file;
@@ -367,13 +364,11 @@ mod tests {
     #[allow(deprecated)]
     #[actix_rt::test]
     async fn status_code_customize_same_output() {
-        let file1 = NamedFile::open_async("Cargo.toml")
-            .await
+        let file1 = NamedFile::open("Cargo.toml")
             .unwrap()
             .set_status_code(StatusCode::NOT_FOUND);
 
-        let file2 = NamedFile::open_async("Cargo.toml")
-            .await
+        let file2 = NamedFile::open("Cargo.toml")
             .unwrap()
             .customize()
             .with_status(StatusCode::NOT_FOUND);
@@ -388,7 +383,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_status_code_text() {
-        let mut file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let mut file = NamedFile::open("Cargo.toml").unwrap();
 
         {
             file.file();
@@ -671,8 +666,7 @@ mod tests {
     async fn test_named_file_content_encoding() {
         let srv = test::init_service(App::new().wrap(Compress::default()).service(
             web::resource("/").to(|| async {
-                NamedFile::open_async("Cargo.toml")
-                    .await
+                NamedFile::open("Cargo.toml")
                     .unwrap()
                     .set_content_encoding(header::ContentEncoding::Identity)
             }),
@@ -693,8 +687,7 @@ mod tests {
     async fn test_named_file_content_encoding_gzip() {
         let srv = test::init_service(App::new().wrap(Compress::default()).service(
             web::resource("/").to(|| async {
-                NamedFile::open_async("Cargo.toml")
-                    .await
+                NamedFile::open("Cargo.toml")
                     .unwrap()
                     .set_content_encoding(header::ContentEncoding::Gzip)
             }),
@@ -720,7 +713,7 @@ mod tests {
     #[actix_rt::test]
     async fn test_named_file_allowed_method() {
         let req = TestRequest::default().method(Method::GET).to_http_request();
-        let file = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let file = NamedFile::open("Cargo.toml").unwrap();
         let resp = file.respond_to(&req);
         assert_eq!(resp.status(), StatusCode::OK);
     }
@@ -1203,7 +1196,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_serve_named_file() {
-        let factory = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let factory = NamedFile::open("Cargo.toml").unwrap();
         let srv = test::init_service(App::new().service(factory)).await;
 
         let req = TestRequest::get().uri("/Cargo.toml").to_request();
@@ -1221,7 +1214,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_serve_named_file_prefix() {
-        let factory = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let factory = NamedFile::open("Cargo.toml").unwrap();
         let srv =
             test::init_service(App::new().service(web::scope("/test").service(factory))).await;
 
@@ -1240,7 +1233,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_default_service() {
-        let factory = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let factory = NamedFile::open("Cargo.toml").unwrap();
         let srv = test::init_service(App::new().default_service(factory)).await;
 
         for route in ["/foobar", "/baz", "/"].iter() {
@@ -1256,7 +1249,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_default_handler_named_file() {
-        let factory = NamedFile::open_async("Cargo.toml").await.unwrap();
+        let factory = NamedFile::open("Cargo.toml").unwrap();
         let st = Files::new("/", ".")
             .default_handler(factory)
             .new_service(())